morning-activity-backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-dotenv==1.1.0
supabase==2.15.3